ref: App components 분리하기
