+++ src/App.jsx
@@ -1,0 +1,45 @@
+import { useState } from 'react';
+import { v4 as uuid } from 'uuid';
+import TodoForm from './Form';
+import EmptyTodos from './EmptyTodos';
+import Todos from './Todos';
+
+export default function App() {
+  const [todoName, setTodoName] = useState('');
+  const [todos, setTodos] = useState([]);
+
+  const isEmpty = todos.length === 0;
+
+  function handleChangeNewTodo({ target: { value } }) {
+    setTodoName(value);
+  }
+
+  function handleSubmitNewTodo(e) {
+    e.preventDefault();
+    setTodos((prevTodos) => [...prevTodos, {
+      id: uuid(),
+      name: todoName,
+      done: false,
+    }]);
+    setTodoName('');
+  }
+
+  function handleClickDeleteTodo(id) {
+    setTodos((prevTodos) => prevTodos.filter((todo) => todo.id !== id));
+  }
+
+  return (
+    <>
+      <h1>To-do</h1>
+      <TodoForm value={todoName} onChange={handleChangeNewTodo} onSubmit={handleSubmitNewTodo} />
+      <p>
+        {isEmpty ? <EmptyTodos /> : (
+          <Todos
+            todos={todos}
+            onDelete={handleClickDeleteTodo}
+          />
+        )}
+      </p>
+    </>
+  );
+}
